requirements.txt
asgiref==3.5.0
Django==3.2.13
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
